add electron-webpack for babel support
