+++ .gitignore
@@ -12,8 +12,8 @@ coverage
 .prettierrc.json
 .vscode
 
-# production
-build
+# build
+dist
 
 # misc
 .DS_Store
