+++ electron-webpack.json
@@ -1,0 +1,11 @@
+{
+  "title": true,
+  "main": {
+    "sourceDirectory": "src/main"
+  },
+  "renderer": {
+    "sourceDirectory": "src/renderer",
+    "template": "src/renderer/index.html"
+  },
+  "commonSourceDirectory": "src/common"
+}
