+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "lights-manager",
   "version": "0.1.0",
-  "main": "public/electron.js",
+  "main": "src/main/main.js",
   "private": true,
   "dependencies": {
     "@material-ui/core": "^4.11.4",
@@ -19,8 +19,8 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "electron": "wait-on tcp:3000 && electron-webpack dev",
-    "dev": "concurrently -k \"BROWSER=none npm start\" \"npm:electron\"",
-    "webpack": "electron-webpack"
+    "dev": "concurrently -k \"BROWSER=none yarn start\" \"yarn electron\"",
+    "webpack": "electron-webpack dev"
   },
   "eslintConfig": {
     "extends": [
@@ -41,16 +41,15 @@
     ]
   },
   "devDependencies": {
-    "@babel/core": "^7.14.6",
-    "@babel/preset-env": "^7.14.7",
     "@babel/preset-react": "^7.14.5",
-    "babel-loader": "^8.1.0",
     "concurrently": "^6.2.0",
     "electron": "^13.1.4",
     "electron-builder": "^22.11.7",
     "electron-is-dev": "^2.0.0",
     "electron-webpack": "^2.8.2",
+    "node-sass": "^6.0.1",
     "react-scripts": "4.0.3",
+    "sass-loader": "^12.1.0",
     "wait-on": "^6.0.0",
     "webpack": "^4.44.2"
   }
