+++ src/common/serial.js
