+++ src/renderer/components/ToggleSwitch.js
