+++ src/renderer/manifest.json
